requirements.txt
chromadb==0.5.3
python-dotenv==1.0.1
requests==2.32.3
sentence-transformers==2.5.1
torch==2.2.1
